vector-store/requirements.txt
flask
faiss-cpu
numpy
requests
